Add support for GCC dwarf2 data
libdrgn/dwarf_info_cache.c uses dwarf_formudata() to pull constants
(like structure member offsets) out of the dwarf data.  Unfortunately,
GCC in -gdwarf-2 mode generates structure constants as blocks with
a single "DW_OP_plus_uconst" operation, which doesn't work with
dwarf_formudata().

This patch introduces a drgn_formudata() wrapper around
dwarf_formudata(), which upon failure sees if the attribute is a
block of that form, and if so returns the proper value.  We have all
the existing dwarf_formudata() callsites use the new wrapper.  This
allows for debugging kernels with dwarf-2 data (which is the default
for x86 linux)