python: update globals even if exescript() raises exception
execscript() is supposed to behave more or less like the script was
typed in at the console. However, if the script raises an exception,
then the variables or functions it defines are not added to the calling
context, which can be pretty confusing.

The problem is that runpy.run_code() returns the new globals, so if it
raises an exception, then we can't get them. Fix it by doing the exec()
ourselves; we just need to do the same setup that runpy does.