helpers: translate task_state_to_char() to Python
Commit 326107f05431 ("libdrgn: add task_state_to_char() helper")
implemented task_state_to_char() in libdrgn so that it could be used in
commit 4780c7a26621 ("libdrgn: stack_trace: prohibit unwinding stack of
running tasks"). As of commit eea542254600 ("libdrgn: make Linux kernel
stack unwinding more robust"), it is no longer used in libdrgn, so we
can translate it to Python. This removes a bunch of code and is more
useful as an example.

Signed-off-by: Omar Sandoval <osandov@osandov.com>