libkdumpfile API usage regression - fix kdump_vmcoreinfo_raw leak
As of b4b3c94cc24dc88d08701ebe35f72a1f1785046a , libkdumpfile
changed some of its API infrastructure and ended up breaking
`kdump_vmcoreinfo_raw` temporarily - The regression was later
fixed with fedbe7f66561754782165be2512bca0ed1b775e9 but the
call now allocates a buffer in `kdump_vmcoreinfo_raw` that
is later up to the caller to free it. This patch updates our
usage of this function so we free our vmcoreinfo string buffer
after we copy the contents that we need from it.

Closes #75.

Signed-off-by: Serapheim Dimitropoulos <serapheim@delphix.com>