Proof-of-concept PEP 517 builder to replace setuptools
setuptools sucks. It's a big pile of accidental behavior that no one
understands. It's designed around monkey-patching, and we hack it in
various ways to make it do what we want. And, since drgn uses GNU
Autotools for the C extension and doesn't have any Python dependencies,
setuptools actually provides us very little value (the same goes for
other packaging tools, including poetry). We just need a way to build
the extension and a way to produce distributions.

Luckily, PEP 517 added a standard way to define a build backend. We just
provide a function to build an sdist and a function to build a wheel.
This is surprisingly easy to do using only the standard library. This
adds a proof-of-concept implementation. Some TODOs remain:

* We may need a shim setup.py for backwards compatibility.
* build.py should provide a proper CLI.
* The setup.py test command should come back as a test.py CLI.
* The Travis and Read the Docs configs need to be updated.
* The documentation needs to be updated.

Also bundled with this change is changing the source distribution to
contain a proper libdrgn distribution (generated by make dist), removing
the need for autotools and gawk when building from an sdist. However,
elfutils needs to be built before creating a distribution, so we'll need
to revisit this once we stop bundling elfutils.